requirements.txt
requests
python_dotenv
GitPython